django/requirements/base.txt
amqp==5.0.5
asgiref==3.3.1
atomicwrites==1.4.0
attrs==20.3.0
billiard==3.6.3.0
celery==5.0.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
Django==3.1.7
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-filter==2.4.0
django-timezone-field==4.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
iniconfig==1.1.1
kombu==5.0.2
mysqlclient==2.0.3
packaging==20.9
pluggy==0.13.1
prompt-toolkit==3.0.17
py==1.10.0
PyJWT==2.0.1
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
vine==5.0.0
wcwidth==0.2.5
uWSGI==2.0.19.1
gunicorn==20.0.4
psycopg2==2.8.6
